Combine method and lucene mappers to EngineFieldMapper
